requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.11.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.9.1
black==20.8b1
bleach==3.1.1
Brotli==1.0.9
bs4==0.0.1
cachelib==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorlover==0.3.0
cryptography==2.9.2
cufflinks==0.17.3
cycler==0.10.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.1
defusedxml==0.6.0
discord.py==1.3.3
distlib==0.3.0
Django==3.0.8
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Session==0.3.2
future==0.18.2
GDAL==2.4.4
html5lib==1.0.1
idna==2.9
imageio==2.9.0
ipykernel==5.1.4
ipyparallel==6.3.0
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jellyfish==0.8.2
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-kernel-gateway==2.4.0
jupyterlab==2.0.0
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lib50==2.0.8
lxml==4.5.0
markdown2==2.3.9
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
multidict==4.7.5
mypy==0.770
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.1.5
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pipx==0.15.1.3
plotly==4.5.4
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyobjc-core==6.2.2
pyobjc-framework-Cocoa==6.2.2
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
regex==2020.10.15
requests==2.23.0
retrying==1.3.3
rope==0.16.0
rumps==0.3.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==2.0.1
spotipy==2.16.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
submit50==3.0.2
sympy==1.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
urllib3==1.25.8
userpath==1.3.0
virtualenv==20.0.17
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
yapf==0.29.0
yarl==1.4.2
